License / Style Fixups
Added Apache 2 license file and updated copyright comment on files.

Various style fixups and documentated exception classes.

Get rid of @SuppressWarnings("null") on entire methods.  The GSON
library isn't @NonNull aware so they have to be somewhere, but I don't
like tagging an entire method.

On methods like getRequiredObjectArray(), in the case of an exception
because one of the methods of the array wasn't the correct type, made
the exception reference the index of the errant entry.