Implement toString on MapCompositeImpl
Added a toString method to call the serializer with String.class.  That
wasn't part of the contract but there was code out there that relied
on that behavior to generate hashes to determine if a configuration
node had changed.

Eventually I'd like to add a getDigest() method since that method
of seeing if a tree has changed is somewhat common.