Remove distribution info / split module-info for tests
Removed distribution block from POM since it's not registered with
Sonatype at this point.  I'm adding in our local repository on the
building line with -DaltDeploymentRepository in the build scripts.

Created alternate module-info for test running, and ignoring primary
module-info in eclipse settings to avoid red flags.

Updated java level to 17 to match the other projects